Controller Comms and Vesc Comms using queues.
