+++ include/controller_comms.h
@@ -72,60 +72,3 @@ bool sendPacketToController(ReasonType reason)
 }
 
 //------------------------------------------------------
-
-#ifndef FSM_H
-#include <Fsm.h>
-#endif
-
-namespace Comms
-{
-  Event lastCommsEvent = EV_NONE;
-
-  State stOffline(
-      [] {
-        commsFsm.printState(OFFLINE);
-        controller_connected = false;
-      },
-      NULL, NULL);
-
-  State stCtrlrOnline(
-      [] {
-        commsFsm.printState(CTRLR_ONLINE);
-        controller_connected = true;
-      },
-      NULL, NULL);
-
-  State stVescOnline(
-      [] {
-        commsFsm.printState(VESC_ONLINE);
-      },
-      NULL, NULL);
-
-  State stCtrlrVescOnline(
-      [] {
-        commsFsm.printState(CTRLR_VESC_ONLINE);
-      },
-      NULL, NULL);
-
-  Fsm fsm(&stOffline);
-
-  //------------------------------------------------------
-  void addTransitions()
-  {
-    // state_offline
-    fsm.add_transition(&stOffline, &stCtrlrOnline, EV_CTRLR_PKT, NULL);
-    fsm.add_transition(&stOffline, &stVescOnline, EV_VESC_SUCCESS, NULL);
-
-    // stCtrlrOnline
-    fsm.add_transition(&stCtrlrOnline, &stOffline, EV_CTRLR_TIMEOUT, NULL);
-    fsm.add_transition(&stCtrlrOnline, &stCtrlrVescOnline, EV_VESC_SUCCESS, NULL);
-
-    // stVescOnline
-    fsm.add_transition(&stVescOnline, &stCtrlrVescOnline, EV_CTRLR_PKT, NULL);
-    fsm.add_transition(&stVescOnline, &stOffline, EV_VESC_FAILED, NULL);
-
-    // stCtrlrVescOnline
-    fsm.add_transition(&stCtrlrVescOnline, &stVescOnline, EV_CTRLR_TIMEOUT, NULL);
-    fsm.add_transition(&stCtrlrVescOnline, &stCtrlrOnline, EV_VESC_FAILED, NULL);
-  }
-} // namespace Comms
